% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bayes_factor}
\alias{bayes_factor}
\title{Compute the Bayes factor between two models.}
\usage{
bayes_factor(samples1, samples2, ...)
}
\arguments{
\item{samples1}{CoalHMM MCMC samples from the first model.}

\item{samples2}{CoalHMM MCMC samples from the second model.}
}
\value{
Bayes factor \eqn{P(D | M1) / P(D | M2)}.
}
\description{
Compute the Bayes factor between two models.
}
\details{
Computes the model likelihood for two different models and returns the Bayes factor comparing the
two models.
}

