% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coalhmm_chain}
\alias{coalhmm_chain}
\title{Creates an object for manipulating CoalHMM MCMC chains}
\usage{
coalhmm_chain(samples)
}
\arguments{
\item{samples}{Samples from a CoalHMM MCMC run.}
}
\value{
The samples wrapped as an object of class "coalhmm_chain".
}
\description{
Creates an object for manipulating CoalHMM MCMC chains
}

