% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convergence_probabilities}
\alias{convergence_probabilities}
\title{Computes the probability of convergence at each windows}
\usage{
convergence_probabilities(rejections, a = 1, b = 1)
}
\description{
Computes the probability, at each window, that the chain was converged at that point, assuming
that the \emph{has} converged at some window.
}
\details{
Computes the probability that convergence happened at a given point based on the pattern of
rejected convergence tests. It assumes that the rejection is picked with a probability that is
beta(a,b) distributed before the change point and rejected with 5% probability after the change
point.
}

