% Generated by roxygen2 (4.0.1): do not edit by hand
\name{equal_distribution}
\alias{equal_distribution}
\title{Compares the distributions for each parameter for two chains to determine if they are equal}
\usage{
equal_distribution(chain1, chain2, strip_chains = TRUE, ...)
}
\arguments{
\item{chain1}{Samples from an MCMC run.}

\item{chain2}{Samples from an MCMC run.}

\item{strip_chains}{Should the chains be stripped of burn-in samples before comparison?}
}
\description{
Compares the samples of two chains to test that for each parameter the samples are from the same distribution.
}
\details{
For each parameter the test does a Kolmogorov-Smirnov test (\code{ks.test}). This is a very simple approach to testing
equality as it doesn't take into account the dependencies between parameters nor that the samples are not independent,
but it is a simple and fast test. It should really be supplemented with visual inspection and other diagnostics.
}

