% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mergeable_chains}
\alias{mergeable_chains}
\title{Tests if a list of chains can be merged}
\usage{
mergeable_chains(chain_list, strip_chains = TRUE, ...)
}
\arguments{
\item{chains}{A \code{list} containing \code{coalhmm_chain}s.}
}
\description{
Test, for all pairs of chains, if they are samples from the same distribution (using \code{equal_distribution}).
}
\details{
The return value contains a p-value computed from another \code{ks.test} testing if the test p-values are uniform.
The idea here is that if each parameter is actually equally distributed then the p-values should be uniform.
}

