% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parameter_samples}
\alias{parameter_samples}
\title{Extract the parameters from the samples data frame}
\usage{
parameter_samples(samples)
}
\arguments{
\item{samples}{Data frame with samples from a CoalHMM MCMC.}
}
\value{
The samples containing only the parameter columns.
}
\description{
Pulls out the columns that corresponds sampled parameters rather than the probabilities,
that is it returns a data frame without the "log.prior", "log.likelihood" and "log.posterior"
columns.
}
\details{
For packages that implements convergence diagnostics, such as coda, we need the samples with
only the parameters and not the prior and posterior probabilities or the likelihood.

This function simply returns that from a data frame of samples as output by the CoalHMM MCMC
samplers.
}

