% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rejected_convergence}
\alias{rejected_convergence}
\title{Determines if the diagnistics means that we reject convergence.}
\usage{
rejected_convergence(convergence_p_values)
}
\arguments{
\item{convergence_p_values}{P-values computed as in \code{convergence_t_tests}.}
}
\value{
A data frame with a row per window specifying if convergence was rejected.
}
\description{
Since the t-test diagnostics give a p-value per parameter we use this function to determine
if the parameters combined leads us to reject convergence. This is simply done by correcting
for multiple tests and then using the smallest p-value per row.
}

